테스트 페이지 삭제 및 develop 브랜치 병합

+++ src/components/team/banner/DeleteTeamModal.tsx
@@ -2,9 +2,10 @@ import { deleteGroupById } from '@/src/api/team/teamAPI';
 import { useTeamStore } from '@/src/stores/teamStore';
 import Button from '@components/@shared/Button';
 import { Modal } from '@components/@shared/Modal';
-import { useMutation } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
+import { UserData } from '../member/ExileUserModal';
 
 interface DeleteTeamModalProps {
   isOpen: boolean;
@@ -17,6 +18,15 @@ export default function DeleteTeamModal({
 }: DeleteTeamModalProps) {
   const { id } = useTeamStore();
   const router = useRouter();
+  const queryClient = useQueryClient();
+  // 'user' 키로 캐싱된 유저 데이터 가져오기
+  const userData = queryClient.getQueryData<UserData>(['user']);
+
+  // 팀 관리자만 삭제 가능
+  const isAdmin =
+    userData &&
+    userData.memberships.find((m) => m.groupId === Number(id))?.role ===
+      'ADMIN';
 
   // 그룹 삭제 Mutation
   const { mutate: deleteGroup } = useMutation({
@@ -26,12 +36,20 @@ export default function DeleteTeamModal({
       onClose();
       router.push('/myteam');
     },
+    onSettled: () => {
+      // 쿼리 무효화 및 리패치
+      queryClient.invalidateQueries({ queryKey: ['user'] });
+    },
     onError: (error) => {
       console.error('팀 삭제 실패:', error);
     },
   });
 
   const handleDeleteClick = () => {
+    // 팀 관리자만 삭제 가능
+    if (!isAdmin) {
+      return;
+    }
     deleteGroup(id as string);
   };
 
@@ -58,12 +76,21 @@ export default function DeleteTeamModal({
             width={24}
             height={24}
           />
-          팀을 삭제하시겠어요?
+          {isAdmin ? (
+            <span>목록을 삭제하시겠어요?</span>
+          ) : (
+            <span>권한 없음</span>
+          )}
         </Modal.Header>
         <Modal.Content fontColor="secondary" fontSize="14" fontArray="center">
-          <p className="mt-[20px]">
-            팀과 관련된 모든 정보가 사라집니다. 정말로 삭제하시겠습니까?
-          </p>
+          {isAdmin && (
+            <p className="mt-[20px]">
+              팀과 관련된 모든 정보가 사라집니다. 정말로 삭제하시겠습니까?
+            </p>
+          )}
+          {!isAdmin && (
+            <p className="mt-[20px]">관리자만 삭제할 수 있습니다.</p>
+          )}
         </Modal.Content>
       </Modal.Wrapper>
       <Modal.Footer>
@@ -76,9 +103,11 @@ export default function DeleteTeamModal({
           >
             취소
           </Button>
-          <Button size="full" bgColor="red" onClick={handleDeleteClick}>
-            삭제
-          </Button>
+          {isAdmin && (
+            <Button size="full" bgColor="red" onClick={handleDeleteClick}>
+              삭제
+            </Button>
+          )}
         </div>
       </Modal.Footer>
     </Modal>
